+++ lib/layrz_models.g.dart
@@ -11,8 +11,8 @@ _$MonitorRealWaypointImpl _$$MonitorRealWaypointImplFromJson(
     _$MonitorRealWaypointImpl(
       activationId: json['activationId'] as String,
       geofenceId: json['geofenceId'] as String,
-      sequenceReal: (json['sequenceReal'] as num?)?.toInt(),
-      sequenceIdeal: (json['sequenceIdeal'] as num).toInt(),
+      sequenceReal: json['sequenceReal'] as int?,
+      sequenceIdeal: json['sequenceIdeal'] as int,
       startAt:
           const TimestampOrNullConverter().fromJson(json['startAt'] as num?),
       endAt: const TimestampOrNullConverter().fromJson(json['endAt'] as num?),
@@ -253,7 +253,7 @@ _$AlgorithmImpl _$$AlgorithmImplFromJson(Map<String, dynamic> json) =>
               ?.map((e) => CredentialField.fromJson(e as Map<String, dynamic>))
               .toList() ??
           const [],
-      usage: (json['usage'] as num?)?.toInt(),
+      usage: json['usage'] as int?,
     );
 
 Map<String, dynamic> _$$AlgorithmImplToJson(_$AlgorithmImpl instance) =>
@@ -555,14 +555,14 @@ _$BillingPlanImpl _$$BillingPlanImplFromJson(Map<String, dynamic> json) =>
     _$BillingPlanImpl(
       id: json['id'] as String,
       reconnectionPercent: (json['reconnectionPercent'] as num).toDouble(),
-      reconnectionMaximum: (json['reconnectionMaximum'] as num).toInt(),
-      reconnectionIncidents: (json['reconnectionIncidents'] as num).toInt(),
-      maxAssets: (json['maxAssets'] as num).toInt(),
-      maxDevices: (json['maxDevices'] as num).toInt(),
-      maxUsers: (json['maxUsers'] as num).toInt(),
-      maxOutboundServices: (json['maxOutboundServices'] as num).toInt(),
-      maxFunctions: (json['maxFunctions'] as num).toInt(),
-      maxApps: (json['maxApps'] as num).toInt(),
+      reconnectionMaximum: json['reconnectionMaximum'] as int,
+      reconnectionIncidents: json['reconnectionIncidents'] as int,
+      maxAssets: json['maxAssets'] as int,
+      maxDevices: json['maxDevices'] as int,
+      maxUsers: json['maxUsers'] as int,
+      maxOutboundServices: json['maxOutboundServices'] as int,
+      maxFunctions: json['maxFunctions'] as int,
+      maxApps: json['maxApps'] as int,
       allowedAppsIds: (json['allowedAppsIds'] as List<dynamic>)
           .map((e) => e as String)
           .toList(),
@@ -625,9 +625,9 @@ _$CareProtocolImpl _$$CareProtocolImplFromJson(Map<String, dynamic> json) =>
       associatedTriggers: (json['associatedTriggers'] as List<dynamic>?)
           ?.map((e) => Trigger.fromJson(e as Map<String, dynamic>))
           .toList(),
-      numOfTasks: (json['numOfTasks'] as num?)?.toInt(),
-      numOfPages: (json['numOfPages'] as num?)?.toInt(),
-      numOfBlocks: (json['numOfBlocks'] as num?)?.toInt(),
+      numOfTasks: json['numOfTasks'] as int?,
+      numOfPages: json['numOfPages'] as int?,
+      numOfBlocks: json['numOfBlocks'] as int?,
       access: (json['access'] as List<dynamic>?)
           ?.map((e) => Access.fromJson(e as Map<String, dynamic>))
           .toList(),
@@ -656,8 +656,8 @@ _$CareTaskImpl _$$CareTaskImplFromJson(Map<String, dynamic> json) =>
               ?.map((e) => e as String)
               .toList() ??
           const [],
-      minValue: (json['minValue'] as num?)?.toInt() ?? 0,
-      maxValue: (json['maxValue'] as num?)?.toInt() ?? 0,
+      minValue: json['minValue'] as int? ?? 0,
+      maxValue: json['maxValue'] as int? ?? 0,
     );
 
 Map<String, dynamic> _$$CareTaskImplToJson(_$CareTaskImpl instance) =>
@@ -714,7 +714,7 @@ _$CaseImpl _$$CaseImplFromJson(Map<String, dynamic> json) => _$CaseImpl(
           .fromJson(json['ignoredStatus'] as String?),
       asset: Asset.fromJson(json['asset'] as Map<String, dynamic>),
       trigger: Trigger.fromJson(json['trigger'] as Map<String, dynamic>),
-      sequence: (json['sequence'] as num?)?.toInt(),
+      sequence: json['sequence'] as int?,
       comments: (json['comments'] as List<dynamic>?)
               ?.map((e) => CaseComment.fromJson(e as Map<String, dynamic>))
               .toList() ??
@@ -869,7 +869,7 @@ _$CloudEntryImpl _$$CloudEntryImplFromJson(Map<String, dynamic> json) =>
       path: json['path'] as String,
       serial: json['serial'] as String?,
       fileId: json['fileId'] as String?,
-      size: (json['size'] as num?)?.toInt(),
+      size: json['size'] as int?,
       lastModified: const TimestampOrNullConverter()
           .fromJson(json['lastModified'] as num?),
       contentType: json['contentType'] as String?,
@@ -1050,8 +1050,8 @@ _$CredentialFieldImpl _$$CredentialFieldImplFromJson(
       field: json['field'] as String,
       type:
           const CredentialFieldTypeConverter().fromJson(json['type'] as String),
-      maxLength: (json['maxLength'] as num?)?.toInt(),
-      minLength: (json['minLength'] as num?)?.toInt(),
+      maxLength: json['maxLength'] as int?,
+      minLength: json['minLength'] as int?,
       maxValue: json['maxValue'] as num?,
       minValue: json['minValue'] as num?,
       choices:
@@ -1105,8 +1105,8 @@ _$CredentialFieldInputImpl _$$CredentialFieldInputImplFromJson(
           ? CredentialFieldType.string
           : const CredentialFieldTypeConverter()
               .fromJson(json['type'] as String),
-      maxLength: (json['maxLength'] as num?)?.toInt(),
-      minLength: (json['minLength'] as num?)?.toInt(),
+      maxLength: json['maxLength'] as int?,
+      minLength: json['minLength'] as int?,
       maxValue: (json['maxValue'] as num?)?.toDouble(),
       minValue: (json['minValue'] as num?)?.toDouble(),
       choices: (json['choices'] as List<dynamic>?)
@@ -1190,7 +1190,7 @@ _$DbPartitionImpl _$$DbPartitionImplFromJson(Map<String, dynamic> json) =>
       startAt:
           const TimestampOrNullConverter().fromJson(json['startAt'] as num?),
       endAt: const TimestampOrNullConverter().fromJson(json['endAt'] as num?),
-      numOfRecords: (json['numOfRecords'] as num?)?.toInt(),
+      numOfRecords: json['numOfRecords'] as int?,
       totalSize: (json['totalSize'] as num?)?.toDouble(),
       sizePerRecord: (json['sizePerRecord'] as num?)?.toDouble(),
     );
@@ -1263,6 +1263,7 @@ _$DeviceImpl _$$DeviceImplFromJson(Map<String, dynamic> json) => _$DeviceImpl(
       modbus: json['modbus'] == null
           ? null
           : ModbusConfig.fromJson(json['modbus'] as Map<String, dynamic>),
+      isSuspended: json['isSuspended'] as bool?,
     );
 
 Map<String, dynamic> _$$DeviceImplToJson(_$DeviceImpl instance) =>
@@ -1285,6 +1286,7 @@ Map<String, dynamic> _$$DeviceImplToJson(_$DeviceImpl instance) =>
       'visionProfile': instance.visionProfile?.toJson(),
       'phone': instance.phone?.toJson(),
       'modbus': instance.modbus?.toJson(),
+      'isSuspended': instance.isSuspended,
     };
 
 _$EmailTemplateImpl _$$EmailTemplateImplFromJson(Map<String, dynamic> json) =>
@@ -1358,7 +1360,7 @@ Map<String, dynamic> _$$EmployeeImplToJson(_$EmployeeImpl instance) =>
 _$FtpAccountImpl _$$FtpAccountImplFromJson(Map<String, dynamic> json) =>
     _$FtpAccountImpl(
       host: json['host'] as String?,
-      port: (json['port'] as num?)?.toInt(),
+      port: json['port'] as int?,
       username: json['username'] as String?,
       password: json['password'] as String?,
     );
@@ -1669,7 +1671,7 @@ Map<String, dynamic> _$$LanguageImplToJson(_$LanguageImpl instance) =>
 _$LintErrorImpl _$$LintErrorImplFromJson(Map<String, dynamic> json) =>
     _$LintErrorImpl(
       code: json['code'] as String,
-      line: (json['line'] as num?)?.toInt() ?? 1,
+      line: json['line'] as int? ?? 1,
       name: json['name'] as String?,
       expected: json['expected'],
       received: json['received'],
@@ -1747,6 +1749,10 @@ _$OperationImpl _$$OperationImplFromJson(Map<String, dynamic> json) =>
       useAssetContactsInstead: json['useAssetContactsInstead'] as bool?,
       attachImage: json['attachImage'] as bool?,
       emailTemplateId: json['emailTemplateId'] as String?,
+      pushPlatforms: (json['pushPlatforms'] as List<dynamic>?)
+          ?.map((e) => const AppPlatformConverter().fromJson(e as String))
+          .toList(),
+      pushTitle: json['pushTitle'] as String?,
     );
 
 Map<String, dynamic> _$$OperationImplToJson(_$OperationImpl instance) =>
@@ -1777,6 +1783,10 @@ Map<String, dynamic> _$$OperationImplToJson(_$OperationImpl instance) =>
       'useAssetContactsInstead': instance.useAssetContactsInstead,
       'attachImage': instance.attachImage,
       'emailTemplateId': instance.emailTemplateId,
+      'pushPlatforms': instance.pushPlatforms
+          ?.map(const AppPlatformConverter().toJson)
+          .toList(),
+      'pushTitle': instance.pushTitle,
     };
 
 _$HttpHeaderImpl _$$HttpHeaderImplFromJson(Map<String, dynamic> json) =>
@@ -2033,7 +2043,7 @@ Map<String, dynamic> _$$CustomReportImplToJson(_$CustomReportImpl instance) =>
 _$SensorImpl _$$SensorImplFromJson(Map<String, dynamic> json) => _$SensorImpl(
       id: json['id'] as String,
       name: json['name'] as String,
-      iterationCycle: (json['iterationCycle'] as num?)?.toInt(),
+      iterationCycle: json['iterationCycle'] as int?,
       slug: json['slug'] as String,
       isInstant: json['isInstant'] as bool?,
       icon: const IconOrNullConverter().fromJson(json['icon'] as String?),
@@ -2455,7 +2465,7 @@ _$TelemetryPositionImpl _$$TelemetryPositionImplFromJson(
       altitude: (json['altitude'] as num?)?.toDouble(),
       speed: (json['speed'] as num?)?.toDouble(),
       direction: (json['direction'] as num?)?.toDouble(),
-      satellites: (json['satellites'] as num?)?.toInt(),
+      satellites: json['satellites'] as int?,
       hdop: (json['hdop'] as num?)?.toDouble(),
       timestamp:
           const TimestampOrNullConverter().fromJson(json['timestamp'] as num?),
@@ -2559,7 +2569,7 @@ _$TriggerImpl _$$TriggerImplFromJson(Map<String, dynamic> json) =>
           .toList(),
       isPlainCrontab: json['isPlainCrontab'] as bool?,
       timezoneId: json['timezoneId'] as String?,
-      priority: (json['priority'] as num?)?.toInt(),
+      priority: json['priority'] as int?,
       color: const ColorOrNullConverter().fromJson(json['color'] as String?),
       visualEventEffect: const CaseEventEffectOrNullConverter()
           .fromJson(json['visualEventEffect'] as String?),
@@ -2710,7 +2720,7 @@ _$TriggerInputImpl _$$TriggerInputImplFromJson(Map<String, dynamic> json) =>
           ],
       isPlainCrontab: json['isPlainCrontab'] as bool? ?? false,
       timezoneId: json['timezoneId'] as String?,
-      priority: (json['priority'] as num?)?.toInt(),
+      priority: json['priority'] as int?,
       color: const ColorOrNullConverter().fromJson(json['color'] as String?),
       visualEventEffect: const CaseEventEffectOrNullConverter()
           .fromJson(json['visualEventEffect'] as String?),
@@ -2853,7 +2863,7 @@ _$UserImpl _$$UserImplFromJson(Map<String, dynamic> json) => _$UserImpl(
       mappitAssets: (json['mappitAssets'] as List<dynamic>?)
           ?.map((e) => Asset.fromJson(e as Map<String, dynamic>))
           .toList(),
-      historicalDaysAllowed: (json['historicalDaysAllowed'] as num?)?.toInt(),
+      historicalDaysAllowed: json['historicalDaysAllowed'] as int?,
       mappitExternalAccountId: json['mappitExternalAccountId'] as String?,
       mappitExternalAccount: json['mappitExternalAccount'] == null
           ? null